requirements.txt
Django==2.0.13
djangorestframework==3.9.4
pytz==2020.1
